+++ files/index.d.ts
@@ -43,10 +43,7 @@ interface KeyValuePair<K, V> extends Array<K | V> {
   1: V;
 }
 
-export interface Lens {
-  <T, U>(obj: T): U;
-  set<T, U>(str: string, obj: T): U;
-}
+export type Lens<S, A> = (functorFactory: (a: A) => Functor<A>) => (s: S) => Functor<S>;
 
 type Arity1Fn = (x: any) => any;
 type Arity2Fn = (x: any, y: any) => any;
@@ -147,12 +144,12 @@ interface SchemaAsync {
   [key: string]: Promise<boolean>;
 }
 
-interface IsValid {
+export interface IsValid {
   input: object;
   schema: Schema;
 }
 
-interface IsValidAsync {
+export interface IsValidAsync {
   input: object;
   schema: Schema | SchemaAsync;
 }
@@ -173,8 +170,6 @@ type ApplyDiffUpdate = {op:'update', path: string, value: any};
 type ApplyDiffAdd = {op:'add', path: string, value: any};
 type ApplyDiffRemove = {op:'remove', path: string};
 type ApplyDiffRule = ApplyDiffUpdate | ApplyDiffAdd | ApplyDiffRemove;
-
-type Resolved<T> = {status: 'fulfilled', value: T} | {status: 'rejected', reason: string|Error}
 
 // API_MARKER
 
@@ -669,7 +664,7 @@ export function compose<TArgs extends any[], R1, R2, R3, R4, R5, R6, R7, TResult
       f2: (a: R1) => R2,
       f1: (...args: TArgs) => R1
   ]
-): (...args: TArgs) => TResult; // fallback overload if number of composed functions greater than 7
+): (...args: TArgs) => TResult;
 export function compose<TArgs extends any[], R1, R2, R3, R4, R5, R6, R7, TResult>(
   f7: (a: R6) => R7,
   f6: (a: R5) => R6,
@@ -851,7 +846,7 @@ Explanation: It decrements a number.
 Example:
 
 ```
-
+const result = R.dec(2) // => 1
 ```
 
 Categories: Number
@@ -1977,7 +1972,7 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function lens<T, U, V>(getter: (s: T) => U, setter: (a: U, s: T) => V): Lens;
+export function lens<S, A>(getter: (s: S) => A, setter: (a: A, s: S) => S): Lens<S, A>;
 
 /*
 Method: lensIndex
@@ -2001,7 +1996,8 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function lensIndex(index: number): Lens;
+export function lensIndex<A>(n: number): Lens<A[], A>;
+export function lensIndex<A extends any[], N extends number>(n: N): Lens<A, A[N]>;
 
 /*
 Method: lensPath
@@ -2029,8 +2025,41 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function lensPath(path: RamdaPath): Lens;
-export function lensPath(path: string): Lens;
+export function lensPath<S, K0 extends keyof S = keyof S>(path: [K0]): Lens<S, S[K0]>;
+export function lensPath<S, K0 extends keyof S = keyof S, K1 extends keyof S[K0] = keyof S[K0]>(
+  path: [K0, K1],
+): Lens<S, S[K0][K1]>;
+export function lensPath<
+  S,
+  K0 extends keyof S = keyof S,
+  K1 extends keyof S[K0] = keyof S[K0],
+  K2 extends keyof S[K0][K1] = keyof S[K0][K1]
+>(path: [K0, K1, K2]): Lens<S, S[K0][K1][K2]>;
+export function lensPath<
+  S,
+  K0 extends keyof S = keyof S,
+  K1 extends keyof S[K0] = keyof S[K0],
+  K2 extends keyof S[K0][K1] = keyof S[K0][K1],
+  K3 extends keyof S[K0][K1][K2] = keyof S[K0][K1][K2]
+>(path: [K0, K1, K2, K3]): Lens<S, S[K0][K1][K2][K3]>;
+export function lensPath<
+  S,
+  K0 extends keyof S = keyof S,
+  K1 extends keyof S[K0] = keyof S[K0],
+  K2 extends keyof S[K0][K1] = keyof S[K0][K1],
+  K3 extends keyof S[K0][K1][K2] = keyof S[K0][K1][K2],
+  K4 extends keyof S[K0][K1][K2][K3] = keyof S[K0][K1][K2][K3]
+>(path: [K0, K1, K2, K3, K4]): Lens<S, S[K0][K1][K2][K3][K4]>;
+export function lensPath<
+  S,
+  K0 extends keyof S = keyof S,
+  K1 extends keyof S[K0] = keyof S[K0],
+  K2 extends keyof S[K0][K1] = keyof S[K0][K1],
+  K3 extends keyof S[K0][K1][K2] = keyof S[K0][K1][K2],
+  K4 extends keyof S[K0][K1][K2][K3] = keyof S[K0][K1][K2][K3],
+  K5 extends keyof S[K0][K1][K2][K3][K4] = keyof S[K0][K1][K2][K3][K4]
+>(path: [K0, K1, K2, K3, K4, K5]): Lens<S, S[K0][K1][K2][K3][K4][K5]>;
+export function lensPath<S = any, A = any>(path: Path): Lens<S, A>;
 
 /*
 Method: lensProp
@@ -2058,10 +2087,7 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function lensProp(prop: string): {
-  <T, U>(obj: T): U;
-  set<T, U, V>(val: T, obj: U): V;
-};
+export function lensProp<S, K extends keyof S = keyof S>(prop: K): Lens<S, S[K]>;
 
 /*
 Method: over
@@ -2137,8 +2163,8 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function view<T, U>(lens: Lens): (target: T) => U;
-export function view<T, U>(lens: Lens, target: T): U;
+export function view<S, A>(lens: Lens<S, A>): (obj: S) => A;
+export function view<S, A>(lens: Lens<S, A>, obj: S): A;
 
 /*
 Method: map
@@ -3688,6 +3714,41 @@ export function sortBy<T>(sortFn: (a: T) => Ord): (list: T[]) => T[];
 export function sortBy(sortFn: (a: any) => Ord): <T>(list: T[]) => T[];
 
 /*
+Method: sortWith
+
+Explanation: 
+
+Example:
+
+```
+const result = R.sortWith([
+    (a, b) => a.a === b.a ? 0 : a.a > b.a ? 1 : -1,
+    (a, b) => a.b === b.b ? 0 : a.b > b.b ? 1 : -1,
+], [
+  {a: 1, b: 2},
+  {a: 2, b: 1},
+  {a: 2, b: 2},
+  {a: 1, b: 1},
+])
+const expected = [
+  {a: 1, b: 1},
+  {a: 1, b: 2},
+  {a: 2, b: 1},
+  {a: 2, b: 2},
+]
+// => `result` is equal to `expected`
+```
+
+Categories: List
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function sortWith<T>(fns: Array<(a: T, b: T) => number>): (list: T[]) => T[];
+export function sortWith<T>(fns: Array<(a: T, b: T) => number>, list: T[]): T[];
+
+/*
 Method: split
 
 Explanation: Curried version of `String.prototype.split`
@@ -4407,9 +4468,9 @@ export function values<T extends object, K extends keyof T>(obj: T): T[K][];
 Method: when
 
 Explanation: It pass `input` to `predicate` function and if the result is `true`, it will return the result of `whenTrueFn(input)`. 
-
 If the `predicate` returns `false`, then it will simply return `input`.
 
+Example:
 ```
 const predicate = x => typeof x === 'number'
 const whenTrueFn = R.add(11)
@@ -5300,6 +5361,8 @@ Explanation:
 Example:
 
 ```
+const result = R.unnest([1, [2], [[3]]])
+// => [1, 2, [3]]
 ```
 
 Categories: List
@@ -5319,6 +5382,12 @@ Explanation:
 Example:
 
 ```
+const result = R.differenceWith(
+  (a, b) => a.x === b.x,
+  [{x: 1}, {x: 2}],
+  [{x: 1}, {x: 3}]
+)
+// => [{x: 2}]
 ```
 
 Categories:
@@ -5343,16 +5412,18 @@ export function differenceWith<T1, T2>(
 /*
 Method: addIndex
 
-Explanation:
+Explanation: 
 
 Example:
 
 ```
+const result = R.addIndex(R.map)((val, idx) => val + idx + 1, [1, 2, 3])
+// => [2, 4, 6]
 ```
 
 Categories:
 
-Notes: TS typings are oversimplified
+Notes:
 
 */
 // @SINGLE_MARKER
@@ -5362,11 +5433,19 @@ export function addIndex(originalFn: any): (fn: any, list: any[]) => any[];
 /*
 Method: ap
 
-Explanation:
+Explanation: It takes a list of functions and a list of values. Then it returns a list of values obtained by applying each function to each value.
 
 Example:
 
 ```
+const result = R.ap(
+  [
+    x => x + 1,
+    x => x + 2,
+  ],
+  [1, 2, 3]
+)
+// => [2, 3, 4, 3, 4, 5]
 ```
 
 Categories:
@@ -5382,7 +5461,7 @@ export function ap<R, A, B>(fn: (r: R, a: A) => B, fn1: (r: R) => A): (r: R) => 
 /*
 Method: addIndexRight
 
-Explanation:
+Explanation: Same as `R.addIndex`, but it will passed indexes are decreasing, instead of increasing.
 
 Example:
 
@@ -5391,7 +5470,7 @@ Example:
 
 Categories:
 
-Notes: TS typings are oversimplified
+Notes:
 
 */
 // @SINGLE_MARKER
@@ -5401,11 +5480,13 @@ export function addIndexRight(originalFn: any): (fn: any, list: any[]) => any[];
 /*
 Method: aperture
 
-Explanation:
+Explanation: It returns a new list, composed of consecutive `n`-tuples from a `list`.
 
 Example:
 
 ```
+const result = R.aperture(2, [1, 2, 3, 4])
+// => [[1, 2], [2, 3], [3, 4]]
 ```
 
 Categories:
@@ -5426,6 +5507,11 @@ Explanation:
 Example:
 
 ```
+const result = R.applyTo(
+  1,
+  x => x + 1
+)
+// => 2
 ```
 
 Categories:
@@ -5445,6 +5531,11 @@ Explanation:
 Example:
 
 ```
+const result = R.sort(
+  R.ascend(x => x),
+  [2, 1]
+)
+// => [1, 2]
 ```
 
 Categories:
@@ -5464,6 +5555,11 @@ Explanation:
 Example:
 
 ```
+R.sort(
+  R.descend(x => x),
+  [1, 2]
+)
+// => [2, 1]
 ```
 
 Categories:
@@ -5483,6 +5579,10 @@ Explanation:
 Example:
 
 ```
+const result = R.binary(
+  (a, b, c) => a + b + c,
+)(1, 2, 3, 4)
+// => 3
 ```
 
 Categories:
@@ -5501,6 +5601,12 @@ Explanation:
 Example:
 
 ```
+const result = R.call(
+  (a, b) => a + b,
+  1,
+  2
+)
+// => 3
 ```
 
 Categories:
@@ -5519,6 +5625,11 @@ Explanation:
 Example:
 
 ```
+const result = R.collectBy(
+  x => x % 2,
+  [1, 2, 3, 4]
+)
+// => [[2, 4], [1, 3]]
 ```
 
 Categories:
@@ -5533,11 +5644,16 @@ export function collectBy<T, K extends PropertyKey>(keyFn: (value: T) => K): (li
 /*
 Method: comparator
 
-Explanation:
+Explanation: It returns a comparator function that can be used in `sort` method.
 
 Example:
 
 ```
+const result = R.sort(
+  R.comparator((a, b) => a.x < b.x),
+  [{x: 2}, {x: 1}]
+)
+// => [{x: 1}, {x: 2}]
 ```
 
 Categories:
@@ -5556,9 +5672,17 @@ Explanation:
 Example:
 
 ```
+const result = R.composeWith(
+  (fn, intermediateResult) => fn(intermediateResult),
+  [
+    R.map(x => x + 1),
+    R.map(x => x * 2),
+  ]
+)([1, 2, 3])
+// => [3, 5, 7]
 ```
 
-Categories:
+Categories: Function
 
 Notes:
 
@@ -5582,6 +5706,8 @@ Explanation:
 Example:
 
 ```
+const result = R.dissocPath(['a', 'b'], {a: {b: 1, c: 2}})
+// => {a: {c: 2}}
 ```
 
 Categories:
@@ -5623,6 +5749,11 @@ Explanation:
 Example:
 
 ```
+const result = R.dropRepeatsBy(
+  Math.abs,
+  [1, -1, 2, 3, -3]
+)
+// => [1, 2, 3]
 ```
 
 Categories:
@@ -5645,6 +5776,8 @@ Explanation:
 Example:
 
 ```
+const result = [R.empty([1,2,3]), R.empty('foo'), R.empty({x: 1, y: 2})]
+// => [[], '', {}]
 ```
 
 Categories:
@@ -5663,6 +5796,8 @@ Explanation:
 Example:
 
 ```
+const result = R.eqBy(Math.abs, 5, -5)
+// => true
 ```
 
 Categories:
@@ -5696,6 +5831,139 @@ Notes:
 // @SINGLE_MARKER
 export function forEachObjIndexed<T>(fn: (value: T[keyof T], key: keyof T, obj: T) => void, obj: T): T;
 export function forEachObjIndexed<T>(fn: (value: T[keyof T], key: keyof T, obj: T) => void): (obj: T) => T;
+
+/*
+Method: gt
+
+Explanation:
+
+Example:
+
+```
+const result = [R.gt(2, 1), R.gt(2, 3)]
+// => [true, false]
+```
+
+Categories: Number
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function gt<T>(x: T): T;
+
+/*
+Method: gte
+
+Explanation:
+
+Example:
+
+```
+const result = [R.gte(2, 1), R.gte(2, 2), R.gte(2, 3)]
+// => [true, true, false]
+```
+
+Categories: Number
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function gte<T>(x: T): T;
+
+/*
+Method: reduceBy
+
+Explanation:
+
+Example:
+
+```
+const result = R.reduceBy(
+  (acc, elem) => acc + elem,
+  0,
+  x => x > 2 ? 'big' : 'small',
+  [1, 2, 3, 4, 5]
+)
+// => { big: 12, small: 3 }
+```
+
+Categories: List
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function reduceBy<T, TResult>(
+  valueFn: (acc: TResult, elem: T) => TResult,
+): (a: TResult, b: (elem: T) => string, c: T[]) => { [index: string]: TResult }
+export function reduceBy<T, TResult>(
+  valueFn: (acc: TResult, elem: T) => TResult,
+  acc: TResult,
+): (a: (elem: T) => string, b: T[]) => { [index: string]: TResult }
+export function reduceBy<T, TResult>(
+  valueFn: (acc: TResult, elem: T) => TResult,
+  acc: TResult,
+  keyFn: (elem: T) => string,
+): (list: T[]) => { [index: string]: TResult };
+export function reduceBy<T, TResult>(
+  valueFn: (acc: TResult, elem: T) => TResult,
+  acc: TResult,
+  keyFn: (elem: T) => string,
+  list: T[],
+): { [index: string]: TResult };
+
+/*
+Method: hasIn
+
+Explanation:
+
+Example:
+
+```
+const result = R.hasIn('a', {a: 1})
+// => true
+```
+
+Categories: String
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function hasIn(searchProperty: string): <T>(obj: T) => boolean;
+export function hasIn<T>(searchProperty: string, obj: T): boolean;
+
+/*
+Method: innerJoin
+
+Explanation: It returns a new list by applying a `predicate` function to all elements of `list1` and `list2` and keeping only these elements where `predicate` returns `true`.
+
+Example:
+
+```
+const list1 = [1, 2, 3, 4, 5]
+const list2 = [4, 5, 6]
+const predicate = (x, y) => x >= y
+const result = R.innerJoin(predicate, list1, list2)
+// => [4, 5]
+```
+
+Categories:
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function innerJoin<T1, T2>(
+  pred: (a: T1, b: T2) => boolean,
+): (list1: T1[], list2: T2[]) => T1[];
+export function innerJoin<T1, T2>(
+  pred: (a: T1, b: T2) => boolean,
+  list1: T1[],
+): (list2: T2[]) => T1[];
+export function innerJoin<T1, T2>(pred: (a: T1, b: T2) => boolean, list1: T1[], list2: T2[]): T1[];
 
 // RAMBDAX_MARKER_START
 
@@ -5873,7 +6141,7 @@ export function composeAsync<TArg, R1, R2, R3, R4, R5, R6, R7, TResult>(
       f2: (a: Awaited<R1>) => R2,
       f1: (a: Awaited<TArg>) => R1
   ]
-): (a: TArg | Promise<TArg>) => TResult; // fallback overload if number of composed functions greater than 7
+): (a: TArg | Promise<TArg>) => TResult;
 export function composeAsync<TArg, R1, R2, R3, R4, R5, R6, R7, TResult>(
   f7: (a: Awaited<R6>) => R7,
   f6: (a: Awaited<R5>) => R6,