+++ source/pluck.spec.js
@@ -4,6 +4,10 @@ test('happy', () => {
   expect(pluck('a')([ { a : 1 }, { a : 2 }, { b : 1 } ])).toEqual([ 1, 2 ])
 })
 
+test('with undefined', () => {
+  expect(pluck(undefined)([ { a : 1 }, { a : 2 }, { b : 1 } ])).toEqual([ ])
+})
+
 test('with number', () => {
   const input = [
     [ 1, 2 ],