+++ source/values.js
@@ -2,6 +2,5 @@ import { type } from './type.js'
 
 export function values(obj){
   if (type(obj) !== 'Object') return []
-
   return Object.values(obj)
 }